feat(core): move generateSuffix from core to app
CRD naming is a consumer concern, not a rendering engine responsibility.
Inlined the parseIsoString dependency to keep the utility self-contained.
