+++ javascript/packages/core/index.tsx
@@ -88,7 +88,6 @@ export { useUrlQueryString } from '#core/hooks/routing/use-url-query-string';
 export * from '#core/utils/object-utils';
 export * from '#core/utils/string-utils';
 export * from '#core/utils/time-utils';
-export { generateSuffix } from '#core/utils/name-utils';
 export { TimeZone } from '#core/types/time-types';
 
 export * from '#core/types/common/studio-types';
